refactor(content): added displaytitles and organised pages
