+++ content/about/resources.en.md
@@ -1,5 +1,6 @@
 ---
-title: Resources
+title: resources
+displayTitle: Resources
 description: 'Resources | Node.js'
 authors: XhmikosR, mikeal, fhemberger, Fishrock123, yous, tomgco, tniessen, SMcCandlish, saadq, Trott, Gornstats, piperchester, naoufal, lpinca, j9t, bnoordhuis, harshadsabne, Chris911, MrJithil
 category: about
