+++ src/pages/about/governance.tsx
@@ -1,36 +1,38 @@
 import { graphql } from 'gatsby';
-import ArticleLayout from '../components/Layout/article';
-import { SideNavBarKeys } from '../components/SideNavBar';
-import connectGraphQlArticle from '../components/connectGraphQlArticle';
+import ArticleLayout from '../../components/Layout/article';
+import { SideNavBarKeys } from '../../components/SideNavBar';
+import connectGraphQlArticle from '../../components/connectGraphQlArticle';
 
 export default connectGraphQlArticle(ArticleLayout, {
-  editPath: 'content/about/security.md',
-  sidenavKey: SideNavBarKeys.security,
+  editPath: 'content/about/governance.md',
+  sidenavKey: SideNavBarKeys.governance,
 });
 
 export const query = graphql`
   query ($locale: String!, $defaultLocale: String!) {
     articleCurrentLanguage: mdx(
-      fields: { slug: { eq: "security" }, locale: { eq: $locale } }
+      fields: { slug: { eq: "governance" }, locale: { eq: $locale } }
     ) {
       body
       tableOfContents
       frontmatter {
         title
         description
+        displayTitle
       }
       fields {
         authors
       }
     }
     articleDefaultLanguage: mdx(
-      fields: { slug: { eq: "security" }, locale: { eq: $defaultLocale } }
+      fields: { slug: { eq: "governance" }, locale: { eq: $defaultLocale } }
     ) {
       body
       tableOfContents
       frontmatter {
         title
         description
+        displayTitle
       }
       fields {
         authors
