+++ src/pages/about/trademark.tsx
@@ -1,36 +1,41 @@
 import { graphql } from 'gatsby';
-import ArticleLayout from '../components/Layout/article';
-import { SideNavBarKeys } from '../components/SideNavBar';
-import connectGraphQlArticle from '../components/connectGraphQlArticle';
+import ArticleLayout from '../../components/Layout/article';
+import { SideNavBarKeys } from '../../components/SideNavBar';
+import connectGraphQlArticle from '../../components/connectGraphQlArticle';
 
 export default connectGraphQlArticle(ArticleLayout, {
-  editPath: 'content/about/about.md',
-  sidenavKey: SideNavBarKeys.about,
+  editPath: 'content/about/trademark.md',
+  sidenavKey: SideNavBarKeys.trademark,
 });
 
 export const query = graphql`
   query ($locale: String!, $defaultLocale: String!) {
     articleCurrentLanguage: mdx(
-      fields: { slug: { eq: "about" }, locale: { eq: $locale } }
+      fields: { slug: { eq: "trademark-policy" }, locale: { eq: $locale } }
     ) {
       body
       tableOfContents
       frontmatter {
         title
         description
+        displayTitle
       }
       fields {
         authors
       }
     }
     articleDefaultLanguage: mdx(
-      fields: { slug: { eq: "about" }, locale: { eq: $defaultLocale } }
+      fields: {
+        slug: { eq: "trademark-policy" }
+        locale: { eq: $defaultLocale }
+      }
     ) {
       body
       tableOfContents
       frontmatter {
         title
         description
+        displayTitle
       }
       fields {
         authors
