+++ src/pages/community.tsx
@@ -20,6 +20,7 @@ export const query = graphql`
       frontmatter {
         title
         description
+        displayTitle
       }
       fields {
         authors
@@ -36,6 +37,7 @@ export const query = graphql`
       frontmatter {
         title
         description
+        displayTitle
       }
       fields {
         authors
